stat: Change access to io_sample union
Compiling on certain platforms produces warnings like

stat.c: In function ‘add_clat_sample’:
stat.c:2241:28: warning: cast from pointer to integer of different size [-Wpointer-to-int-cast]

because the size of a pointer is different to that of a long. Work
around this by explicitly naming the union used in io_sample, forcing
the "val" in the union to be a uint64_t, changing function parameters to
accept a union (rather than an unsigned long) and adding helper
functions to wrap values into a union.

Signed-off-by: Sitsofe Wheeler <sitsofe@yahoo.com>